requirements.txt
keras
tensorflow-gpu
opencv-python
matplotlib
pandas
